feat: subgraphQueries files changed to ponderQueries and new query added to the file. Removed the three previous tables to use only one
